* int to quint * Martingale default settings modified

+++ casinos/spingenerator/spingeneratorsettings.cpp
@@ -2,12 +2,12 @@
 
 const QString STARTING_CASH = "StartingCash";
 
-const int SpinGeneratorSettings::GetStartingCash() const
+const quint8 SpinGeneratorSettings::GetStartingCash() const
 {
-	return _csSettings.value(STARTING_CASH, 10).toInt();
+	return _csSettings.value(STARTING_CASH, 10).toUInt();
 } // GetStartingCash
 
-const void SpinGeneratorSettings::SetStartingCash(const int &pCash)
+const void SpinGeneratorSettings::SetStartingCash(const quint8 &pCash)
 {
 	_csSettings.setValue(STARTING_CASH, pCash);
 } // SetStartingCash
