+++ systems/martingale/martingalestatisticswidget.h
@@ -26,7 +26,7 @@ class MartingaleStatisticsWidget : public QWidget
 
 		const void Increment(const eCounter &pCounter) const;
 		const void Reset() const;
-		const void SetMaxSameInRow(const int &pValue) const;
+		const void SetMaxSameInRow(const quint8 &pValue) const;
 
 	private:
 		Ui::qwMartingaleStatisticsWidget _qwmswWidget;
